+++ DynamicFestivalRewards/ModConfig.cs
@@ -1,0 +1,8 @@
+﻿namespace DynamicFestivalRewards;
+
+internal class ModConfig
+{
+    public bool Randomize { get; set; } = true;
+    public int MinValue { get; set; } = 200;
+    public int MaxValue { get; set; } = 1000;
+}